third_party/WebKit/LayoutTests/fast/scrolling/hover-during-scroll-expected.txt

was hovered
was hovered
currently hovered
hover over me
hover over me
hover over me
hover over me
hover over me
hover over me
hover over me
hover over me
hover over me
hover over me
hover over me
hover over me
hover over me
hover over me
hover over me
hover over me
hover over me
hover over me
hover over me
hover over me
hover over me
hover over me
hover over me
hover over me
hover over me
hover over me
hover over me
hover over me
Test for http://crbug.com/153784. New hover effects should not be invoked during scroll if the mouse cursor is not visible.

PASS successfullyParsed is true

TEST COMPLETE
Mouse is visible, moving it over the first div.
PASS array[0].innerHTML is "currently hovered"
PASS array[1].innerHTML is "hover over me"
Mouse is visible, moving it over the second div.
PASS array[0].innerHTML is "was hovered"
PASS array[1].innerHTML is "currently hovered"
PASS array[2].innerHTML is "hover over me"
Mouse is visible, scrolling page so the mouse ends up on the third div.
PASS document.scrollingElement.scrollTop is 0
PASS document.scrollingElement.scrollTop == elementHeight became true
PASS array[0].innerHTML is "was hovered"
PASS array[1].innerHTML is "was hovered"
PASS array[2].innerHTML is "currently hovered"
PASS array[3].innerHTML is "hover over me"
Mouse is not visible, scrolling page so the mouse ends up on the fourth div.
PASS document.scrollingElement.scrollTop is 50
PASS document.scrollingElement.scrollTop == elementHeight*2 became true
PASS array[0].innerHTML is "was hovered"
PASS array[1].innerHTML is "was hovered"
PASS array[2].innerHTML is "currently hovered"
PASS array[3].innerHTML is "hover over me"
PASS array[4].innerHTML is "hover over me"
PASS document.scrollingElement.scrollTop is 100
